refactoring: rebuild spec structure, added constants, updated wdio.conf.js, added beforeEach hooks, added missing prettier config files, added example of env file, removed logger pino, removed unnecessary files, updated README.md instruction
