+++  .eslintignore
@@ -1,3 +1,4 @@
 node_modules/
 allure-report/
 allure-results/
+wdio.conf.js
