+++ .env.example
@@ -1,0 +1,6 @@
+DEVICE_NAME=
+PLATFORM_VERSION=
+AUTOMATION_DRIVER_NAME=
+LOG_LEVEL=info
+RETRIES=1
+TEST_USERNAME=
