+++ .gitignore
@@ -1,0 +1,6 @@
+/node_modules
+/logs
+.env
+package-lock.json
+/allure-report
+/allure-results
