+++ .prettierignore
@@ -1,0 +1,3 @@
+node_modules/
+allure-report/
+allure-results/
