+++ package.json
@@ -3,13 +3,15 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "wdio:run": "wdio run ./wdio.conf.js",
-    "appium:run": "appium --port 3000 --address localhost --allow-insecure chromedriver_autodownload,adb_shell",
+    "allure:clear": "rimraf allure-results allure-report",
     "allure:generate": "npx allure generate allure-results --clean -o allure-report",
+    "appium:run": "appium --port 3000 --address localhost",
     "allure:open": "npx allure open allure-report",
     "allure:generate:open": "npm run allure:generate && npm run allure:open",
     "lint": "eslint ./src ./tests --ext .js",
-    "prettier": "prettier --write './**/*.{js,json,md}'"
+    "lint: fix": "eslint ./src ./tests --ext .js --fix",
+    "prettier": "prettier --write './**/*.{js,json,md}'",
+    "wdio:run": "npm run allure:clear && wdio run ./wdio.conf.js"
   },
   "author": "",
   "license": "ISC",
@@ -31,12 +33,20 @@
     "mocha": "^11.1.0",
     "pino-pretty": "^13.0.0",
     "prettier": "^3.5.3",
+    "rimraf": "^6.0.1",
     "wait-port": "^1.1.0",
+    "wdio-simple-reporter": "^0.1.2",
     "wdio-wait-for": "^3.1.0"
   },
   "dependencies": {
     "dotenv": "^16.5.0",
     "jimp": "^0.16.1",
+    "module-alias": "^2.2.3",
     "pino": "^9.6.0"
+  },
+  "_moduleAliases": {
+    "#pages": "src/pages",
+    "#commands": "src/commands",
+    "#utils": "src/utils"
   }
 }
