--- src/scripts/clearReporter.sh
@@ -1,3 +1,0 @@
-rm -rf ../allure-report ../allure-results
-echo "Allure cleared."
-exit 0
