+++ src/utils/constants.js
@@ -1,0 +1,16 @@
+const constants = {
+  CART: {
+    TITLE: 'My Cart',
+    EMPTY_CART_TEXT: 'No Items'
+  },
+  LOGIN: {
+    USERNAME_REQUIRED_ERROR: 'Username is required',
+    PASSWORD_REQUIRED_ERROR: 'Enter Password'
+  },
+  PRODUCT: {
+    PRODUCT_TITLE: 'Products',
+    SUCCESS_MESSAGE: 'Thank you for submitting your review!'
+  }
+};
+
+module.exports = constants;
